% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_mass.R
\name{approximate_pmf_kolmogorov}
\alias{approximate_pmf_kolmogorov}
\title{Use Kolmogorov to approximate the pmf of a binomial convolution distribution}
\usage{
approximate_pmf_kolmogorov(n_trials, success_probs)
}
\arguments{
\item{n_trials}{A vector of positive integers each standing for the number of trials of one component}

\item{success_probs}{A vector of floats between 0 and 1 each standing for the probability of success of one component}
}
\value{
The pmf of the specified binomial convolution distribution
}
\description{
This function uses Kolmogorov approximation to compute the pmf of a specified
binomial convolution distribution
}
