% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_functions.R
\name{binomial_cf}
\alias{binomial_cf}
\title{Characteristic function of a binomial distribution}
\usage{
binomial_cf(n_trials, success_prob)
}
\arguments{
\item{n_trials}{The number of trials}

\item{success_prob}{The probability of success}
}
\value{
The CF of the specified binomial distribution
}
\description{
This function returns the CF of a binomial distribution
given the number of trials and probability of success
}
