% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_functions.R
\name{binomial_convolution_pmf}
\alias{binomial_convolution_pmf}
\title{The pmf of a binomial convolution distribution}
\usage{
binomial_convolution_pmf(
  n_trials,
  success_probs,
  computation_method = "convolution"
)
}
\arguments{
\item{n_trials}{A vector of positive integers each standing for the number of trials of one component}

\item{success_probs}{A vector of floats between 0 and 1 each standing for the probability of success of one component}

\item{computation_method}{The method to use to compute the pmf}
}
\value{
The pmf of the specified binomial convolution distribution
}
\description{
This function computes the pmf of a specified
binomial convolution distribution
}
