% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_functions.R
\name{binomial_cumulants}
\alias{binomial_cumulants}
\title{The first six cumulants of a binomial distribution}
\usage{
binomial_cumulants(n_trials, success_prob)
}
\arguments{
\item{n_trials}{A positive integer standing for the number of trials}

\item{success_prob}{A floats between 0 and 1 standing for the probability of success}
}
\value{
The first six cumulants of the specified binomial distribution
}
\description{
This function returns the first six cumulants of a
binomial distribution given the number of trials
and the success probability
}
