% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_gf.R
\name{binomial_mgf}
\alias{binomial_mgf}
\title{Moment generating function of a binomial distribution}
\usage{
binomial_mgf(n_trials, success_prob)
}
\arguments{
\item{n_trials}{The number of trials}

\item{success_prob}{The probability of success}
}
\value{
The MGF of the specified binomial distribution
}
\description{
This function returns the MGF of a binomial distribution
given the number of trials and probability of success
}
