% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_functions.R
\name{central_moments}
\alias{central_moments}
\title{The first six central moments of a binomial convolution distribution}
\usage{
central_moments(n_trials, success_probs)
}
\arguments{
\item{n_trials}{A vector of positive integers each standing for the number of trials of one component}

\item{success_probs}{A vector of floats between 0 and 1 each standing for the probability of success of one component}
}
\value{
The first six central moments of the specified binomial convolution distribution
}
\description{
This function returns the first six central moments of a
binomial covolution distribution given the number of trials
and the success probabilities
}
