% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_gf.R
\name{compute_discrepancy}
\alias{compute_discrepancy}
\title{Compute discrepancy}
\usage{
compute_discrepancy(integrand, lower_limit, upper_limit)
}
\arguments{
\item{integrand}{The integrand function}

\item{lower_limit}{The lower bound of the integral}

\item{upper_limit}{The upper bound of the integral}
}
\value{
The discrepancy value
}
\description{
This function performs numerical integration given
the integrand as well as the limits of the integral
}
