% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{cumulants}
\alias{cumulants}
\title{The first six cumulants of a binomial convolution distribution}
\usage{
cumulants(n_trials, success_probs)
}
\arguments{
\item{n_trials}{A vector of positive integers each standing for the number of trials of one component}

\item{success_probs}{A vector of floats between 0 and 1 each standing for the probability of success of one component}
}
\value{
The first six cumulants of the specified binomial convolution distribution
}
\description{
This function returns the first six cumulants of a
binomial covolution distribution given the number of trials
and the success probabilities
}
