% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_gf.R
\name{d_binomial_cgf}
\alias{d_binomial_cgf}
\title{First order derivative of the cumulant generating function of a binomial distribution}
\usage{
d_binomial_cgf(n_trials, success_prob)
}
\arguments{
\item{n_trials}{The number of trials}

\item{success_prob}{The probability of success}
}
\value{
The derivative of the CGF of the specified binomial distribution
}
\description{
This function returns the derivative of the CGF of a binomial distribution
given the number of trials and probability of success
}
