% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{dbinomconv}
\alias{dbinomconv}
\title{Density of a Binomial Convolution}
\usage{
dbinomconv(x, n_trials, success_probs, log = FALSE)
}
\arguments{
\item{x}{Vector of quantiles}

\item{n_trials}{A vector of positive integers each standing for the number of trials of one component}

\item{success_probs}{A vector of floats between 0 and 1 each standing for the probability of success of one component}

\item{n_samples}{Number of samples to be generated}
}
\value{
A vector of probabilities
}
\description{
Density of a Binomial Convolution
}
