% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_mass.R
\name{dft_helper}
\alias{dft_helper}
\title{Computes some key quantities in the DFT}
\usage{
dft_helper(n_trials, success_probs)
}
\arguments{
\item{n_trials}{A vector of positive integers each standing for the number of trials of one component}

\item{success_probs}{A vector of floats between 0 and 1 each standing for the probability of success of one component}
}
\value{
A list of quantities used in DFT computation
}
\description{
This function computes some key quantities used repeatedly
in the DFT computation.
}
