% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mle2.R
\name{diff_chisq}
\alias{diff_chisq}
\title{Given a radius and an angle, compute the difference between the chisq stats}
\usage{
diff_chisq(
  r,
  theta,
  solution_to_use,
  solution_nllh,
  n_trials,
  samples,
  sig_level = 0.95
)
}
\arguments{
\item{r}{The radius}

\item{theta}{The angle}

\item{solution_to_use}{The MLE solution}

\item{solution_nllh}{The negative log likelihood at the MLE}

\item{n_trials}{The number of trials for each component}

\item{samples}{A sample from the BC distribution}

\item{sig_level}{The significance level}
}
\value{
The difference between chisq stats
}
\description{
This function is used for computing the joint confidence region based on
the likelihood ratio test.
}
