% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_gf.R
\name{discrepancy_integrand}
\alias{discrepancy_integrand}
\title{The integrand of the final discrepancy function}
\usage{
discrepancy_integrand(theoretical_gf, empirical_gf, weight_function)
}
\arguments{
\item{theoretical_gf}{The theoretical generating function (MGF OR CGF)}

\item{empirical_gf}{The empirical estimator of the gf (MGF OR CGF)}

\item{weight_function}{A weighting function that ensures the convergence of the integral}
}
\value{
The integrand of the discrepancy function
}
\description{
This function returns the integrand of the discrepancy function
given the theoretical generating function, its empirical
estimator, and a weight function
}
