% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_gf.R
\name{empirical_cf}
\alias{empirical_cf}
\title{Empirical estimates of the CF of a Binomial Convolution distribution}
\usage{
empirical_cf(samples)
}
\arguments{
\item{samples}{A vector of IID samples}
}
\value{
The ECF of a binomial convolution distribution
}
\description{
This function returns an empirical estimator of the CF of
a binomial convolution distribution given a set of samples
}
