% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_functions.R
\name{empirical_cgf}
\alias{empirical_cgf}
\title{Empirical estimates of the CGF of a Binomial Convolution distribution}
\usage{
empirical_cgf(samples)
}
\arguments{
\item{samples}{A vector of IID samples}
}
\value{
The ECGF of a binomial convolution distribution
}
\description{
This function returns an empirical estimator of the CGF of
a binomial convolution distribution given a set of samples
}
