% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_functions.R
\name{empirical_mgf}
\alias{empirical_mgf}
\title{Empirical estimates of the MGF of a Binomial Convolution distribution}
\usage{
empirical_mgf(samples)
}
\arguments{
\item{samples}{A vector of IID samples}
}
\value{
The EMGF of a binomial convolution distribution
}
\description{
This function returns an empirical estimator of the MGF of
a binomial convolution distribution given a set of samples
}
