% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_functions.R
\name{exact_pmf_convolution}
\alias{exact_pmf_convolution}
\title{Use convolution to compute the exact pmf of a binomial convolution distribution}
\usage{
exact_pmf_convolution(n_trials, success_probs)
}
\arguments{
\item{n_trials}{A vector of positive integers each standing for the number of trials of one component}

\item{success_probs}{A vector of floats between 0 and 1 each standing for the probability of success of one component}
}
\value{
The pmf of the specified binomial convolution distribution
}
\description{
This function uses convolution to compute the pmf of a specified
binomial convolution distribution
}
