% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is_legit_solution}
\alias{is_legit_solution}
\title{Check if a vector contains all probabilities}
\usage{
is_legit_solution(est)
}
\arguments{
\item{est}{A vector}
}
\value{
If all elements are probabilities
}
\description{
Check if a vector contains all probabilities
}
