% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mle_general.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Compute log likelihood}
\usage{
log_likelihood(
  success_probs,
  n_trials,
  samples,
  computation_method = "convolution",
  return_mean = FALSE
)
}
\arguments{
\item{success_probs}{A vector of floats between 0 and 1 each standing for the probability of success of one component}

\item{n_trials}{A vector of positive integers each standing for the number of trials of one component}

\item{samples}{A vector of samples from a binomial convolution distribution}

\item{computation_method}{The method to use to compute the pmf}

\item{return_mean}{Whether or not to return average instead of the sum}
}
\value{
the log likelihood
}
\description{
This function computes the log likelihood given
a sample as well as the parameters
}
