% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mle_general.R
\name{log_likelihood_equality_constraint}
\alias{log_likelihood_equality_constraint}
\title{Log likelihood that includes some fixed probabilities}
\usage{
log_likelihood_equality_constraint(
  par,
  n_trials,
  samples,
  fixed_success_probs,
  fixed_success_probs_index,
  computation_method = "convolution",
  return_mean = FALSE
)
}
\arguments{
\item{par}{Logits of non-fixed probabilities}

\item{n_trials}{The number of trials for each component}

\item{samples}{A sample from the BC distribution}

\item{fixed_success_probs}{Probabilities of the fixed components}

\item{fixed_success_probs_index}{The corresponding indecies of fixed_success_probs}

\item{computation_method}{The method to compute the PMF}

\item{return_mean}{If the results should be averaged over the sample}
}
\value{
The log likelihood
}
\description{
To compute profile likelihood, a subset of the components
is fixed. This function facilitates this exact purpose.
}
