% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mle_general.R
\name{log_likelihood_order_constraint}
\alias{log_likelihood_order_constraint}
\title{Computes the log likelihood with possible order constraint}
\usage{
log_likelihood_order_constraint(
  par,
  n_trials,
  samples,
  computation_method = "convolution",
  return_mean = FALSE,
  order_success_probs = NULL
)
}
\arguments{
\item{par}{The parameters (not the probabilities)}

\item{n_trials}{The number of trials for each component}

\item{samples}{A sample from the BC distribution}

\item{computation_method}{The method to compute the PMF}

\item{return_mean}{If the results should be averaged over the sample}

\item{order_success_probs}{A vector specifying the order in descending order of probabilities}
}
\value{
The log likelihood
}
\description{
This function computes the log likelihood with order constraint
}
