% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make_pmf}
\alias{make_pmf}
\title{Converts a vector into a pmf}
\usage{
make_pmf(pmf_vector)
}
\arguments{
\item{pmf_vector}{A vector of probabilities}
}
\value{
A pmf whose support starts at 0
}
\description{
This function takes a vector and turns it into
a pmf whose support starts at 0
}
