% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{mle}
\alias{mle}
\title{MLE of a given sample from a Binomial Convolution distribution}
\usage{
mle(
  samples,
  n_trials,
  computation_method = "convolution",
  return_mean = FALSE,
  order_success_probs = NULL,
  initial_probs = NULL
)
}
\arguments{
\item{samples}{A sample from the BC distribution}

\item{n_trials}{The number of trials for each component}

\item{computation_method}{The method to compute the PMF}

\item{return_mean}{If the results should be averaged over the sample}

\item{order_success_probs}{A vector specifying the order in descending order of probabilities}

\item{initial_probs}{The initial guess of the probabilies}
}
\value{
The MLE and the corresponding log likelihood
}
\description{
This function computes the MLE of a BC model given a sample and
the number of trials for each component. It also allows the users
to specify the order of probabilities as well as initial guesses.
}
