% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mle2.R
\name{mle2}
\alias{mle2}
\title{Computes MLE for a 2-component BC model}
\usage{
mle2(
  samples,
  n_trials,
  computation_method = "convolution",
  return_mean = FALSE,
  initial_probs = NULL,
  report_both = FALSE
)
}
\arguments{
\item{samples}{A sample from the BC distribution}

\item{n_trials}{The number of trials for each component}

\item{computation_method}{The method to compute the PMF}

\item{return_mean}{If the results should be averaged over the sample}

\item{initial_probs}{The initial guess of the probabilities}

\item{report_both}{If report both solutions}
}
\value{
The MLE estimates as well as their corresponding log likelihood
}
\description{
Due to the pecularity of a BC distribution, we single out the 2-component
model as the simplies case.
}
