% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mom2.R
\name{mom_2}
\alias{mom_2}
\title{Method of moment for a 2-component Binomial Convolution model}
\usage{
mom_2(samples, n_trials)
}
\arguments{
\item{samples}{A vector of samples generated from a binomial convolution model}

\item{n_trials}{A vector specifying the number of trials}
}
\value{
A vector of estimated probabilities
}
\description{
This function performs method of moment for a 2-component model
}
