% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mom2.R
\name{mom_2_helper}
\alias{mom_2_helper}
\title{Helper function for method of moment for a 2-component Binomial Convolution model}
\usage{
mom_2_helper(samples, n_trials)
}
\arguments{
\item{samples}{A vector of samples generated from a binomial convolution model}

\item{n_trials}{A vector specifying the number of trials}
}
\value{
A list of two solutions to the mom.
}
\description{
This function performs method of moment for a 2-component model
}
