% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{par_to_probs}
\alias{par_to_probs}
\title{Converts a vector of parameters into probabilities}
\usage{
par_to_probs(par, order_success_probs = NULL)
}
\arguments{
\item{par}{A vector of parameters}

\item{order_success_probs}{A vector specifying the order in descending order of probabilities}
}
\value{
The transformed probabilities of the vector of parameters
}
\description{
As a constraint optimization problem is in general more difficult
than an unconstraint one, we convert probabilities to numbers on the real line.
This function further considers the order. If there is no order imposed,
the probabilities are computed via the inverse logit function.
On the other hand, the first parameter will be transformed via the inverse logit.
The rest will be converted via the cumulative sum of exp of difference in logits and reordered.
}
