% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prob_fence}
\alias{prob_fence}
\title{Truncate a vector so that everything falls between the boundaries}
\usage{
prob_fence(vec, UB = 1, LB = 0)
}
\arguments{
\item{vec}{A vector}

\item{UB}{The upper bound}

\item{LB}{The lower bound}
}
\value{
The trunccated vector
}
\description{
Truncate a vector so that everything falls between the boundaries
}
