% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{probs_to_par}
\alias{probs_to_par}
\title{Converts a vector of probabilities into parameters}
\usage{
probs_to_par(probs, order_success_probs = NULL)
}
\arguments{
\item{probs}{A vector of probabilities}

\item{order_success_probs}{A vector specifying the order in descending order of probabilities}
}
\value{
The transformed parameters of the vector of probabilities
}
\description{
As a constraint optimization problem is in general more difficult
than an unconstraint one, we convert probabilities to numbers on the real line.
This function further considers the order. If there is no order imposed,
the probabilities are converted via the logit function.
On the other hand, the largest probability will be transformed via the logit.
The rest will be converted via the log of difference in logits.
}
