% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mle2.R
\name{profile_likelihood_chisq}
\alias{profile_likelihood_chisq}
\title{Compute the chisq stat based on profile likelihood}
\usage{
profile_likelihood_chisq(
  n_trials,
  samples,
  fixed_success_probs,
  fixed_success_probs_index
)
}
\arguments{
\item{n_trials}{The number of trials for each component}

\item{samples}{A sample from the BC distribution}

\item{fixed_success_probs}{Probabilities of the fixed components}

\item{fixed_success_probs_index}{The corresponding indecies of fixed_success_probs}
}
\value{
The Chisq statistics
}
\description{
Compute the chisq stat based on profile likelihood
}
