% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mle_general.R
\name{profile_log_likelihood}
\alias{profile_log_likelihood}
\title{Profile log likelihood of a Binomial Convolution model}
\usage{
profile_log_likelihood(
  samples,
  n_trials,
  fixed_success_probs,
  fixed_success_probs_index,
  computation_method = "convolution",
  return_mean = FALSE,
  initial_probs = NULL
)
}
\arguments{
\item{samples}{A sample from the BC distribution}

\item{n_trials}{The number of trials for each component}

\item{fixed_success_probs}{Probabilities of the fixed components}

\item{fixed_success_probs_index}{The corresponding indecies of fixed_success_probs}

\item{computation_method}{The method to compute the PMF}

\item{return_mean}{If the results should be averaged over the sample}

\item{initial_probs}{The initial guess of the non-fixed probabilities}
}
\value{
The MLE and the corresponding log likelihood
}
\description{
Computes the MLE when a subset of the parameters are fixed
This method is useful when constructing marginal confidence intervals/regions
}
