% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generating_functions.R
\name{theoretical_cf}
\alias{theoretical_cf}
\title{Characteristic function of a Binomial Convolution distribution}
\usage{
theoretical_cf(n_trials, success_probs)
}
\arguments{
\item{n_trials}{A vector of positive integers each standing for the number of trials of one component}

\item{success_probs}{A vector of floats between 0 and 1 each standing for the probability of success of one component}
}
\value{
The CF of the specified binomial convolution distribution
}
\description{
This function returns the CF of a Binomial Convolution distribution
given the number of trials and the success probabilities
}
