% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mle2.R
\name{trace_confidence_region}
\alias{trace_confidence_region}
\title{Use points to trace out the contour of the confidence region}
\usage{
trace_confidence_region(solution_to_use, solution_nllh, n_trials, samples)
}
\arguments{
\item{solution_to_use}{The MLE solution}

\item{solution_nllh}{The negative log likelihood at the MLE}

\item{n_trials}{The number of trials for each component}

\item{samples}{A sample from the BC distribution}
}
\value{
The contour traced out by points
}
\description{
Use points to trace out the contour of the confidence region
}
