% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mle2.R
\name{trace_contour}
\alias{trace_contour}
\title{Given two points, sample points in the middle to trace out the contour of the confidence region}
\usage{
trace_contour(
  theta1,
  theta2,
  r1,
  r2,
  solution_to_use,
  solution_nllh,
  n_trials,
  samples,
  pots = c()
)
}
\arguments{
\item{theta1}{The angle of the first point}

\item{theta2}{The angle of the second point}

\item{r1}{The radius of the first point}

\item{r2}{The radius of the second point}

\item{solution_to_use}{The MLE solution}

\item{solution_nllh}{The negative log likelihood at the MLE}

\item{n_trials}{The number of trials for each component}

\item{samples}{A sample from the BC distribution}
}
\value{
The contour traced out by points between the given two points
}
\description{
Given two points, sample points in the middle to trace out the contour of the confidence region
}
