% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{v_last_which}
\alias{v_last_which}
\title{Returns the index of the last element that is greater than p}
\usage{
v_last_which(x, p)
}
\arguments{
\item{x}{A vector}

\item{p}{A vector}
}
\value{
The radius and the angle
}
\description{
Returns the index of the last element that is greater than p
}
