% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mle2.R
\name{vec_diff_chisq}
\alias{vec_diff_chisq}
\title{Vectorized version of diff_chisq}
\usage{
vec_diff_chisq(
  r,
  theta,
  solution_to_use,
  solution_nllh,
  n_trials,
  samples,
  sig_level = 0.95
)
}
\description{
Vectorized version of diff_chisq
}
