% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_mle2.R
\name{vector_to_theta_r}
\alias{vector_to_theta_r}
\title{Transform a R2 vector from Cartisean coordinate to polar coordinate}
\usage{
vector_to_theta_r(vec)
}
\arguments{
\item{vec}{A vector}
}
\value{
The radius and the angle
}
\description{
Transform a R2 vector from Cartisean coordinate to polar coordinate
}
